+++ contracts/fungible-tokens/andromeda-cw20-exchange/src/testing/tests.rs
@@ -404,7 +404,6 @@ pub fn test_purchase_not_enough_sent() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(100u128),
         },
     )
     .unwrap();
@@ -453,7 +452,6 @@ pub fn test_purchase_no_tokens_left() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::zero(),
         },
     )
     .unwrap();
@@ -496,7 +494,6 @@ pub fn test_purchase_not_enough_tokens() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::one(),
         },
     )
     .unwrap();
@@ -540,7 +537,6 @@ pub fn test_purchase() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -628,7 +624,6 @@ pub fn test_purchase_with_start_and_duration() {
             start_time: Expiration::AtTime(env.block.time.minus_nanos(1)),
             // end time in the future
             end_time: Expiration::AtTime(env.block.time.plus_nanos(1)),
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -715,7 +710,6 @@ pub fn test_purchase_sale_not_started() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time.plus_nanos(1)),
             end_time: Expiration::Never {},
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -760,7 +754,6 @@ pub fn test_purchase_sale_duration_ended() {
             start_time: Expiration::AtTime(env.block.time),
 
             end_time: Expiration::AtTime(env.block.time.minus_nanos(1)),
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -818,7 +811,6 @@ pub fn test_purchase_not_enough_sent_native() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(100u128),
         },
     )
     .unwrap();
@@ -858,7 +850,6 @@ pub fn test_purchase_no_tokens_left_native() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::zero(),
         },
     )
     .unwrap();
@@ -894,7 +885,6 @@ pub fn test_purchase_not_enough_tokens_native() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(1u128),
         },
     )
     .unwrap();
@@ -932,7 +922,6 @@ pub fn test_purchase_native() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -1011,7 +1000,6 @@ pub fn test_purchase_refund() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(100u128),
         },
     )
     .unwrap();
@@ -1060,7 +1048,6 @@ pub fn test_cancel_sale_unauthorised() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -1118,7 +1105,6 @@ pub fn test_cancel_sale() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: sale_amount,
         },
     )
     .unwrap();
@@ -1179,7 +1165,6 @@ fn test_query_sale() {
         start_time: Expiration::AtTime(env.block.time),
 
         end_time: Expiration::Never {},
-        start_amount: sale_amount,
     };
     SALE.save(deps.as_mut().storage, &exchange_asset.to_string(), &sale)
         .unwrap();
@@ -1219,7 +1204,6 @@ fn test_andr_query() {
         start_time: Expiration::AtTime(env.block.time),
 
         end_time: Expiration::Never {},
-        start_amount: sale_amount,
     };
     SALE.save(deps.as_mut().storage, &exchange_asset.to_string(), &sale)
         .unwrap();
@@ -1260,7 +1244,6 @@ fn test_purchase_native_invalid_coins() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(100u128),
         },
     )
     .unwrap();
@@ -1313,7 +1296,6 @@ fn test_query_sale_assets() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(100u128),
         },
     )
     .unwrap();
@@ -1326,7 +1308,6 @@ fn test_query_sale_assets() {
             recipient: owner.to_string(),
             start_time: Expiration::AtTime(env.block.time),
             end_time: Expiration::Never {},
-            start_amount: Uint128::from(100u128),
         },
     )
     .unwrap();