+++ packages/std/src/error.rs
@@ -601,6 +601,9 @@ pub enum ContractError {
     #[error("Sale not ended")]
     SaleNotEnded {},
 
+    #[error("Redemption condition already exists")]
+    RedemptionConditionAlreadyExists {},
+
     #[error("Min sales exceeded")]
     MinSalesExceeded {},